Created new separate folder for UpdateJoke
